feat: removed api service and made the service as it's own files
